docsitalia: introduce the docs italia publisher hierarchy
This adds a separate app called docsitalia to handle the
required Publisher -> Project -> Document hierarchy.

On top of this we need a custom DocsItaliaGitHubService
in order to parse our custom metadata for publishers and
projects and act accordingly.

To use it you should add to your local settings:

CLASS_OVERRIDES = {
   'readthedocs.oauth.services.GitHubService':
       'docsitalia.oauth.services.github.DocsItaliaGithubService'
}