search: avoid elastic search returning hits with very low score
For some reason ES returns hits with score 0 in the file search.
This goes unnoticed when the hits return something sensible but
in the case the query does not match anything we don't want to
throw random stuff at the user. So we add a completely arbitrary
min_score of 1 to avoid unwanted hits.