docsitalia: don't enforce slug and name uniqueness for PublisherProject
As we don't want it :) First we need the slug to be unique only for
the same publisher. Then it would pretty legitimate to have the same
project name (and thus slug) for different publisher.

After auditing PublisherProject querysets the only change behaviour
would be in DocsItaliaProjectViewSet where we let filtering
the queryset by publisher project slug. Since the api permits to
filter for the publisher slug too users that want unique results
need to filter for that too.